requirements.txt
certifi==2020.12.5
jedi==0.17.0
pandas==1.2.3
Pillow==8.2.0
pytz==2021.1
pywin32==227
pyzmq==20.0.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing-extensions==3.7.4.3
wincertstore==0.2
